Módulo_1-Automação-Excel-e-Engenharia-de-Dados/9_automacao_de_processos_com_excel_em_python/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
lxml==6.0.0
numpy==2.3.1
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
wheel==0.45.1
